fix: move get... methods to separate fetchet.ts file
