+++ test-data/shared/article.generator.ts
@@ -1,5 +1,6 @@
+import { getArticlesJSON } from './article.fetcher';
 import { faker } from '@faker-js/faker/locale/en';
-import { APIRequestContext, APIResponse } from '@playwright/test';
+import { APIRequestContext } from '@playwright/test';
 
 // generate unique article ID
 export async function generateUniqueArticleId(
@@ -8,7 +9,7 @@ export async function generateUniqueArticleId(
   maxRange: number = 2000,
 ): Promise<number> {
   // get list of all article IDs
-  const articlesJSON = await generateArticlesJSON(request);
+  const articlesJSON = await getArticlesJSON(request);
   const articleIdList: number[] = articlesJSON.map(
     (article: { id: number }) => article.id,
   );
@@ -28,24 +29,11 @@ export async function generateUniqueArticleId(
   return uniqueArticleId;
 }
 
-export async function takeMaxArticleId(
-  request: APIRequestContext,
-): Promise<number> {
-  // get list of all article IDs
-  const articlesJSON = await generateArticlesJSON(request);
-  const articleIdList: number[] = articlesJSON.map(
-    (article: { id: number }) => article.id,
-  );
-
-  // return maximum ID
-  return Math.max(...articleIdList);
-}
-
 export async function generateUniqueArticleTitle(
   request: APIRequestContext,
 ): Promise<string> {
   // get list of all article titles
-  const articlesJSON = await generateArticlesJSON(request);
+  const articlesJSON = await getArticlesJSON(request);
   const articleTitleList = new Set(
     articlesJSON.map((article: { title: string }) => article.title),
   );
@@ -57,21 +45,4 @@ export async function generateUniqueArticleTitle(
   } while (articleTitleList.has(uniqueArticleTitle));
 
   return uniqueArticleTitle;
-}
-
-export async function getExistingArticleTitle(
-  request: APIRequestContext,
-): Promise<string> {
-  const articlesJSON = await generateArticlesJSON(request);
-  const firstArticleTitle: string = articlesJSON[0].title;
-  return firstArticleTitle;
-}
-
-// function to get articles
-export async function generateArticlesJSON(
-  request: APIRequestContext,
-): Promise<{ id: number; title: string }[]> {
-  const getAllArticles: APIResponse = await request.get(`/api/articles`);
-  const articlesJSON = await getAllArticles.json();
-  return articlesJSON;
 }
